# internal/app/rest_api_v2/docs/swagger.yaml
basePath: /api/v2
definitions:
  FreeBSDOsInfo.ArcInfo:
    properties:
      arc_used_bytes:
        type: integer
      arc_used_human:
        type: string
    type: object
  FreeBSDOsInfo.CpuInfo:
    properties:
      cpu_arch:
        type: string
      cpu_cores:
        type: integer
      cpu_model:
        type: string
      cpu_sockets:
        type: integer
      cpu_threads:
        type: integer
      overall_cpus:
        type: integer
    type: object
  FreeBSDOsInfo.IoStatCpu:
    properties:
      idle_mode:
        description: '% of cpu time in idle mode'
        type: integer
      system_interrupt_mode:
        description: '% of cpu time in system interrupt mode'
        type: integer
      system_mode:
        description: '% of cpu time in system mode'
        type: integer
      user_mode:
        description: '% of cpu time in user mode'
        type: integer
      user_nice_mode:
        description: '% of cpu time in user nice mode'
        type: integer
    type: object
  FreeBSDOsInfo.RamInfo:
    properties:
      ram_free_bytes:
        type: integer
      ram_free_human:
        type: string
      ram_overall_bytes:
        type: integer
      ram_overall_human:
        type: string
      ram_used_bytes:
        type: integer
      ram_used_human:
        type: string
    type: object
  FreeBSDOsInfo.SwapInfo:
    properties:
      swap_free_bytes:
        type: integer
      swap_free_human:
        type: string
      swap_overall_bytes:
        type: integer
      swap_overall_human:
        type: string
      swap_used_bytes:
        type: integer
      swap_used_human:
        type: string
    type: object
...
  HosterHostUtils.HostInfo:
    properties:
      all_vms:
        type: integer
      arc_info:
        $ref: '#/definitions/FreeBSDOsInfo.ArcInfo'
      backup_vms:
        type: integer
      cpu_info:
        $ref: '#/definitions/FreeBSDOsInfo.CpuInfo'
      cpu_metrics:
        $ref: '#/definitions/FreeBSDOsInfo.IoStatCpu'
      hostname:
        type: string
      live_vms:
        type: integer
      offline_vms:
        type: integer
      offline_vms_prod:
        type: integer
      ram_info:
        $ref: '#/definitions/FreeBSDOsInfo.RamInfo'
      services:
        $ref: '#/definitions/HosterHostUtils.HosterServices'
      swap_info:
        $ref: '#/definitions/FreeBSDOsInfo.SwapInfo'
      system_major_version:
        type: string
      system_uptime:
        type: string
      vcpu_2_pcpu_ratio:
        type: number
      zpool_list:
        items:
          $ref: '#/definitions/zfsutils.ZpoolInfo'
        type: array
    type: object
  HosterHostUtils.HosterServices:
    properties:
      dns_server_pid:
        type: integer
      dns_server_running:
        type: boolean
      dns_server_version:
        type: string
      ha_watchdog_pid:
        type: integer
      ha_watchdog_running:
        type: boolean
      ha_watchdog_version:
        type: string
      hoster_version:
        type: string
      mbuffer_version:
        type: string
      node_exporter_pid:
        type: integer
      node_exporter_running:
        type: boolean
      node_exporter_version:
        type: string
      rest_api_pid:
        type: integer
      rest_api_running:
        type: boolean
      rest_api_version:
        type: string
      scheduler_pid:
        type: integer
      scheduler_running:
        type: boolean
      scheduler_version:
        type: string
      self_update_version:
        type: string
      vm_supervisor_version:
        type: string
    type: object
...
  HosterVm.VmDeployInput:
    properties:
      custom_dns_server:
        type: string
      ip_address:
        type: string
      network_name:
        type: string
      os_type:
        type: string
      ram:
        type: string
      start_when_ready:
        type: boolean
      target_dataset:
        type: string
      vcpus:
        type: integer
      vm_name:
        type: string
    type: object
  HosterVmUtils.Virtio9P:
    properties:
      read_only:
        type: boolean
      share_location:
        type: string
      share_name:
        type: string
    type: object
  HosterVmUtils.VmApi:
    properties:
      9p_shares:
        items:
          $ref: '#/definitions/HosterVmUtils.Virtio9P'
        type: array
      backup:
        type: boolean
      cpu_cores:
        type: integer
      cpu_sockets:
        type: integer
      cpu_threads:
        type: integer
      current_host:
        type: string
      description:
        type: string
      disable_xhci:
        type: boolean
      disks:
        items:
          $ref: '#/definitions/HosterVmUtils.VmDisk'
        type: array
      dns_search_domain:
        type: string
      encrypted:
        type: boolean
      include_host_ssh_keys:
        type: boolean
      loader:
        type: string
      memory:
        type: string
      name:
        type: string
      networks:
        items:
          $ref: '#/definitions/HosterVmUtils.VmNetwork'
        type: array
      os_comment:
        type: string
      os_type:
        type: string
      owner:
        type: string
      parent_host:
        type: string
      passthru:
        items:
          type: string
        type: array
      production:
        description: LiveStatus         string      `json:"live_status"`
        type: boolean
      running:
        type: boolean
      tags:
        items:
          type: string
        type: array
      uptime:
        type: string
      uptime_unix:
        type: integer
      uuid:
        type: string
      vga:
        type: string
      vm_ssh_keys:
        items:
          $ref: '#/definitions/HosterVmUtils.VmSshKey'
        type: array
      vnc_password:
        type: string
      vnc_port:
        type: integer
      vnc_resolution:
        type: integer
    type: object
  HosterVmUtils.VmDisk:
    properties:
      comment:
        type: string
      disk_image:
        type: string
      disk_location:
        type: string
      disk_type:
        type: string
      total_bytes:
        type: integer
      total_human:
        type: string
      used_bytes:
        type: integer
      used_human:
        type: string
    type: object
  HosterVmUtils.VmNetwork:
    properties:
      comment:
        type: string
      ip_address:
        type: string
      network_adaptor_type:
        type: string
      network_bridge:
        type: string
      network_mac:
        type: string
    type: object
  HosterVmUtils.VmSshKey:
    properties:
      comment:
        type: string
      key_owner:
        type: string
      key_value:
        type: string
    type: object
...
  handlers.SwaggerError:
    properties:
      id:
        type: integer
      message:
        type: string
    type: object
  handlers.SwaggerSuccess:
    properties:
      message:
        description: success
        type: string
    type: object
  handlers.VmCloneInput:
    properties:
      new_vm_name:
        type: string
      snapshot_name:
        type: string
      vm_name:
        type: string
    type: object
  handlers.VmStopInput:
    properties:
      force_cleanup:
        description: Kill the VM supervisor directly (useful in the situations where
          you want to destroy the VM, or roll it back to a previous snapshot)
        type: boolean
      force_stop:
        description: Send a SIGKILL instead of a graceful SIGTERM
        type: boolean
      vm_name:
        type: string
    type: object
...
  zfsutils.ZpoolInfo:
    properties:
      allocated_bytes:
        type: integer
      allocated_human:
        type: string
      fragmentation:
        type: integer
      free_bytes:
        type: integer
      free_human:
        type: string
      healthy:
        type: boolean
      name:
        type: string
      size_bytes:
        type: integer
      size_human:
        type: string
    type: object
info:
  contact: {}
  description: '`NOTE!` This REST API HTTP endpoint is located directly on the `Hoster`
    node.<br><br>The API should ideally be integrated into another system (e.g. a
    user-accessible back-end server), and not interacted with directly.<br><br>Please,
    take an extra care with the things you execute here, because some of them may
    be disruptive or non-revertible (e.g. vm destroy, snapshot rollback, host reboot,
    etc).'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Hoster Node REST API Docs
  version: "2.0"
paths:
...
  /host/info:
    get:
      description: 'Get Host info.<br>`AUTH`: Both users are allowed.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HosterHostUtils.HostInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Get Host info.
      tags:
      - Host
...
  /vm/all:
    get:
      description: 'Get the list of all VMs, including the information about them.<br>`AUTH`:
        Both users are allowed.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/HosterVmUtils.VmApi'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: List all VMs.
      tags:
      - VMs
  /vm/clone:
    post:
      description: 'Clone the VM using it''s name, and optionally specify the snapshot
        name to be used for cloning.<br>`AUTH`: Only `rest` user is allowed.'
      parameters:
      - description: Request payload
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/handlers.VmCloneInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Clone the VM.
      tags:
      - VMs
  /vm/deploy:
    post:
      description: 'Deploy a new VM.<br>`AUTH`: Only `rest` user is allowed.'
      parameters:
      - description: Request payload
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/HosterVm.VmDeployInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Deploy the new VM.
      tags:
      - VMs
  /vm/destroy/{vm_name}:
    delete:
      description: 'Destroy the VM using it''s name.<br>`AUTH`: Only `rest` user is
        allowed.'
      parameters:
      - description: VM Name
        in: path
        name: vm_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Destroy the VM.
      tags:
      - VMs
  /vm/info/{vm_name}:
    get:
      description: 'Get the VM Info.<br>`AUTH`: Both users are allowed.'
      parameters:
      - description: VM Name
        in: path
        name: vm_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HosterVmUtils.VmApi'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Get the VM Info.
      tags:
      - VMs
  /vm/start/{vm_name}:
    post:
      description: 'Start a specific VM using it''s name as a parameter.<br>`AUTH`:
        Both users are allowed.'
      parameters:
      - description: VM Name
        in: path
        name: vm_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Start a specific VM.
      tags:
      - VMs
  /vm/stop:
    post:
      description: 'Stop a specific VM using it''s name as a parameter.<br>`AUTH`:
        Both users are allowed.'
      parameters:
      - description: Request payload
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/handlers.VmStopInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SwaggerSuccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.SwaggerError'
      security:
      - BasicAuth: []
      summary: Stop a specific VM.
      tags:
      - VMs
swagger: "2.0"